+++ instrumentation/opentelemetry-instrumentation-fastapi/tests/test_fastapi_instrumentation.py
@@ -597,17 +597,17 @@ class TestHTTPAppWithCustomHeaders(TestBase):
 
     def test_http_custom_request_headers_in_span_attributes(self):
         expected = {
-            "http.request.header.custom_test_header_1": (
+            "http.request.header.custom-test-header-1": (
                 "test-header-value-1",
             ),
-            "http.request.header.custom_test_header_2": (
+            "http.request.header.custom-test-header-2": (
                 "test-header-value-2",
             ),
-            "http.request.header.regex_test_header_1": ("Regex Test Value 1",),
-            "http.request.header.regex_test_header_2": (
+            "http.request.header.regex-test-header-1": ("Regex Test Value 1",),
+            "http.request.header.regex-test-header-2": (
                 "RegexTestValue2,RegexTestValue3",
             ),
-            "http.request.header.my_secret_header": ("[REDACTED]",),
+            "http.request.header.my-secret-header": ("[REDACTED]",),
         }
         resp = self.client.get(
             "/foobar",
@@ -631,7 +631,7 @@ class TestHTTPAppWithCustomHeaders(TestBase):
 
     def test_http_custom_request_headers_not_in_span_attributes(self):
         not_expected = {
-            "http.request.header.custom_test_header_3": (
+            "http.request.header.custom-test-header-3": (
                 "test-header-value-3",
             ),
         }
@@ -658,19 +658,19 @@ class TestHTTPAppWithCustomHeaders(TestBase):
 
     def test_http_custom_response_headers_in_span_attributes(self):
         expected = {
-            "http.response.header.custom_test_header_1": (
+            "http.response.header.custom-test-header-1": (
                 "test-header-value-1",
             ),
-            "http.response.header.custom_test_header_2": (
+            "http.response.header.custom-test-header-2": (
                 "test-header-value-2",
             ),
-            "http.response.header.my_custom_regex_header_1": (
+            "http.response.header.my-custom-regex-header-1": (
                 "my-custom-regex-value-1,my-custom-regex-value-2",
             ),
-            "http.response.header.my_custom_regex_header_2": (
+            "http.response.header.my-custom-regex-header-2": (
                 "my-custom-regex-value-3,my-custom-regex-value-4",
             ),
-            "http.response.header.my_secret_header": ("[REDACTED]",),
+            "http.response.header.my-secret-header": ("[REDACTED]",),
         }
         resp = self.client.get("/foobar")
         self.assertEqual(200, resp.status_code)
@@ -684,7 +684,7 @@ class TestHTTPAppWithCustomHeaders(TestBase):
 
     def test_http_custom_response_headers_not_in_span_attributes(self):
         not_expected = {
-            "http.response.header.custom_test_header_3": (
+            "http.response.header.custom-test-header-3": (
                 "test-header-value-3",
             ),
         }
@@ -753,10 +753,10 @@ class TestWebSocketAppWithCustomHeaders(TestBase):
 
     def test_web_socket_custom_request_headers_in_span_attributes(self):
         expected = {
-            "http.request.header.custom_test_header_1": (
+            "http.request.header.custom-test-header-1": (
                 "test-header-value-1",
             ),
-            "http.request.header.custom_test_header_2": (
+            "http.request.header.custom-test-header-2": (
                 "test-header-value-2",
             ),
         }
@@ -789,7 +789,7 @@ class TestWebSocketAppWithCustomHeaders(TestBase):
     )
     def test_web_socket_custom_request_headers_not_in_span_attributes(self):
         not_expected = {
-            "http.request.header.custom_test_header_3": (
+            "http.request.header.custom-test-header-3": (
                 "test-header-value-3",
             ),
         }
@@ -816,10 +816,10 @@ class TestWebSocketAppWithCustomHeaders(TestBase):
 
     def test_web_socket_custom_response_headers_in_span_attributes(self):
         expected = {
-            "http.response.header.custom_test_header_1": (
+            "http.response.header.custom-test-header-1": (
                 "test-header-value-1",
             ),
-            "http.response.header.custom_test_header_2": (
+            "http.response.header.custom-test-header-2": (
                 "test-header-value-2",
             ),
         }
@@ -839,7 +839,7 @@ class TestWebSocketAppWithCustomHeaders(TestBase):
 
     def test_web_socket_custom_response_headers_not_in_span_attributes(self):
         not_expected = {
-            "http.response.header.custom_test_header_3": (
+            "http.response.header.custom-test-header-3": (
                 "test-header-value-3",
             ),
         }